+++ Runtime/Extensions/StreamExtensions.cs
@@ -6,6 +6,11 @@ namespace Backtrace.Unity.Extensions
     internal static class StreamExtensions
     {
 #if !(NET_STANDARD_2_0 && NET_4_6)
+        // We pick a value that is the largest multiple of 4096 that is still smaller than the large object heap threshold (85K).
+        // The CopyTo/CopyToAsync buffer is short-lived and is likely to be collected at Gen0, and it offers a significant
+        // improvement in Copy performance.
+        private const int _DefaultCopyBufferSize = 81920;
+
         public static void CopyTo(this Stream original, Stream destination)
         {
             if (destination == null)
@@ -29,7 +34,7 @@ namespace Backtrace.Unity.Extensions
                 throw new NotSupportedException("NotSupportedException destination");
             }
 
-            byte[] array = new byte[4096];
+            byte[] array = new byte[_DefaultCopyBufferSize];
             int count;
             while ((count = original.Read(array, 0, array.Length)) != 0)
             {